pyproject.toml
[tool.poetry]
name = "mdcx"
version = "0.1.0"
description = "Seamless markdown to docx converter."
authors = ["Owen Griffiths <root@ogriffiths.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^0.8.11"
pillow = "^10.0.1"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
